+++ framework/servlet/src/main/java/is/codion/framework/servlet/EntityService.java
@@ -391,7 +391,7 @@ public final class EntityService implements AuxiliaryServer {
 				RemoteEntityConnection connection = authenticate(context);
 				context.status(HttpStatus.OK_200)
 								.contentType(ContentType.APPLICATION_OCTET_STREAM)
-								.result(serialize(connection.isQueryCacheEnabled()));
+								.result(serialize(connection.queryCache()));
 			}
 			catch (Exception e) {
 				handleException(context, e);
@@ -404,7 +404,7 @@ public final class EntityService implements AuxiliaryServer {
 				ObjectMapper objectMapper = objectMapper(connection.entities());
 				context.status(HttpStatus.OK_200)
 								.contentType(ContentType.APPLICATION_JSON)
-								.result(objectMapper.writeValueAsString(connection.isQueryCacheEnabled()));
+								.result(objectMapper.writeValueAsString(connection.queryCache()));
 			}
 			catch (Exception e) {
 				handleException(context, e);
@@ -417,7 +417,7 @@ public final class EntityService implements AuxiliaryServer {
 		private void serial(Context context) {
 			try {
 				RemoteEntityConnection connection = authenticate(context);
-				connection.setQueryCacheEnabled(deserialize(context.req()));
+				connection.queryCache(deserialize(context.req()));
 				context.status(HttpStatus.OK_200);
 			}
 			catch (Exception e) {
@@ -429,7 +429,7 @@ public final class EntityService implements AuxiliaryServer {
 			try {
 				RemoteEntityConnection connection = authenticate(context);
 				ObjectMapper objectMapper = objectMapper(connection.entities());
-				connection.setQueryCacheEnabled(objectMapper.readValue(context.req().getInputStream(), Boolean.class));
+				connection.queryCache(objectMapper.readValue(context.req().getInputStream(), Boolean.class));
 				context.status(HttpStatus.OK_200);
 			}
 			catch (Exception e) {