+++ changelog.md
@@ -26,6 +26,7 @@ Codion Change Log
 - DerivedAttributeDefinition.valueProvider() renamed provider().
 ### is.codion.framework.db.core
 - EntityConnectionProvider.Builder.domainType() renamed domain().
+- EntityConnection.setQueryCacheEnabled() and isQueryCacheEnabled() renamed queryCache().
 ### is.codion.framework.model
 - EntityQueryModel bug fixed, did support including non-selected columns.
 ### is.codion.swing.framework.ui